Added table class and methods changed
