+++ admin/src/main/scala/net/wiringbits/components/pages/TablePage.scala
@@ -1,24 +1,17 @@
 package net.wiringbits.components.pages
 
-import net.wiringbits.components.widgets.Table
-import net.wiringbits.ui.components.core.widgets.Title
+import net.wiringbits.API
+import net.wiringbits.components.widgets.TableWidget
 import slinky.core.FunctionalComponent
 import slinky.core.annotations.react
 import slinky.core.facade.Fragment
-import typings.reactRouter.mod.useParams
-
-import scala.scalajs.js
 
 @react object TablePage {
-  type Props = Unit
+  case class Props(api: API)
 
   val component: FunctionalComponent[Props] = FunctionalComponent[Props] { props =>
-    val params = useParams()
-    val table_name = params.asInstanceOf[js.Dynamic].table_name.toString
-
     Fragment(
-      Title(table_name),
-      Table()
+      TableWidget(props.api)
     )
   }
